requirements.txt
pyyaml
setuptools
numpy
matplotlib
seaborn
box
joblib
ensure
-e .